- Deleted obsolete groundstation code
- Committed current revision of satellite code, less ADC, PanTilt+MPU and SPI data logging logic

+++ firmware/satellite/system/include/stm32f4xx/stm32f4xx_hal_crc.h
@@ -153,7 +153,7 @@ void HAL_CRC_MspDeInit(CRC_HandleTypeDef *hcrc);
   * @{
   */
 uint32_t HAL_CRC_Accumulate(CRC_HandleTypeDef *hcrc, uint32_t pBuffer[], uint32_t BufferLength);
-uint32_t HAL_CRC_Calculate(CRC_HandleTypeDef *hcrc, uint8_t pBuffer[], uint32_t BufferLength);
+uint32_t HAL_CRC_Calculate(CRC_HandleTypeDef *hcrc, uint32_t pBuffer[], uint32_t BufferLength);
 /**
   * @}
   */
