+++ firmware/satellite/system/src/stm32f4xx/stm32f4xx_hal_crc.c
@@ -266,7 +266,7 @@ uint32_t HAL_CRC_Accumulate(CRC_HandleTypeDef *hcrc, uint32_t pBuffer[], uint32_
   * @param  BufferLength: Length of the buffer to be computed
   * @retval 32-bit CRC
   */
-uint32_t HAL_CRC_Calculate(CRC_HandleTypeDef *hcrc, uint8_t pBuffer[], uint32_t BufferLength)
+uint32_t HAL_CRC_Calculate(CRC_HandleTypeDef *hcrc, uint32_t pBuffer[], uint32_t BufferLength)
 {
   uint32_t index = 0U;
